backend/requirements.txt
Django==5.1.6
djangorestframework
django-cors-headers
openpyxl
Pillow

# Optional/dev
# pytest-django
# django-extensions
